Laser Tower, Tower State Setting, ( + Build, Sell Function Complete )

+++ MoluDefence/Assets/Scripts/Projectile.cs
@@ -7,9 +7,9 @@ public class Projectile : MonoBehaviour
 {
     private Movement2D _movement2D;
     private Transform  _target;
-    private int        _damage;
+    private float      _damage;
 
-    public void Setup(Transform target, int damage)
+    public void Setup(Transform target, float damage)
     {
         _movement2D  = GetComponent<Movement2D>();
         this._target = target;
